feature: resetting the votes now works
