+++ internal/service/vote_service.go
@@ -10,15 +10,20 @@ type VoteService interface {
 	Reset(breakout *domain.Breakout)
 }
 
-type voteService struct{}
+type voteService struct {
+	broadcast BroadcastService
+}
 
 func NewVoteService() VoteService {
-	return voteService{}
+	return voteService{
+		broadcast: NewBroadcastService(),
+	}
 }
 
 func (v voteService) Reset(breakout *domain.Breakout) {
 	database.DB.Model(domain.User{}).Where("breakout_id = ?", breakout.ID).Update("vote", "")
 	database.DB.Model(domain.Breakout{}).Where("id = ?", breakout.ID).Update("show_votes", false)
+	v.broadcast.ResetVotes(breakout.ID)
 }
 
 func (v voteService) Vote(user *domain.User, value string) {
@@ -32,6 +37,7 @@ func (v voteService) Vote(user *domain.User, value string) {
 	var breakout domain.Breakout
 	database.DB.Preload("Users", "is_online = ?", true).First(&breakout, "id = ?", user.BreakoutID)
 	database.DB.Model(domain.Breakout{}).Where("id = ?", breakout.ID).Update("show_votes", v.didEveryoneVote(breakout))
+	v.broadcast.Breakout(breakout.ID)
 }
 
 func (v voteService) didEveryoneVote(breakout domain.Breakout) bool {
