+++ templates/breakout/index.html
@@ -17,6 +17,7 @@
         hx-ext="ws"
         ws-connect="/ws/{{ .Breakout.ID }}?user_id={{ .UserID }}"
       >
+        <div id="voting-cards" hx-swap-oob="beforeend">{{ template "vote/index" .Breakout }}</div>
         <div id="breakout" hx-swap-oob="beforeend"></div>
       </div>
       <input id="user_id" type="hidden" value="{{ .UserID }}" />
